+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## 0.2.0
 
+- #10 adds a `deps-new` template that sets up a basic Clerk project with
+  `Mafs.cljs` installed as a dependency. [The template lives
+  here](https://github.com/mentat-collective/Mafs.cljs/tree/main/resources/mafs/clerk).
+
 - #7:
 
   - Renames the `mafs` namespace to `mafs.core` to prevent collisions with the